内容管理

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/xuecheng/content/model/dto/AddCourseDto.java
@@ -10,9 +10,6 @@ import java.math.BigDecimal;
 
 /**
  * @description 添加课程dto
- * @author Mr.M
- * @date 2022/9/7 17:40
- * @version 1.0
  */
 @Data
 @ApiModel(value="AddCourseDto", description="新增课程基本信息")
@@ -56,9 +53,10 @@ public class AddCourseDto {
  private String charge;
 
  @ApiModelProperty(value = "价格")
- private Float price;
+ private BigDecimal price;
+
  @ApiModelProperty(value = "原价")
- private Float originalPrice;
+ private BigDecimal originalPrice;
 
 
  @ApiModelProperty(value = "qq")
